cluster/requirements.txt
numpy==1.25.2
hdbscan==0.8.33
requests==2.31.0
pandas==2.1.1
umap-learn==0.5.4
pika==1.3.2
PyYAML==6.0.1
SQLAlchemy==1.4.49
mysql-connector-python==8.1.0